refactor: 딜러와 플레이어에서 발생하는 중복 코드 제거
