+++ src/main/java/model/blackjackgame/BlackjackGame.java
@@ -3,10 +3,10 @@ package model.blackjackgame;
 import java.util.List;
 import model.card.Card;
 import model.card.Cards;
-import model.dealer.Dealer;
-import model.player.Player;
-import model.player.Players;
-import model.player.UpdatedPlayer;
+import model.participants.dealer.Dealer;
+import model.participants.player.Player;
+import model.participants.player.Players;
+import model.participants.player.UpdatedPlayer;
 
 public class BlackjackGame {
