+++ src/main/java/model/blackjackgame/Profit.java
@@ -2,8 +2,8 @@ package model.blackjackgame;
 
 import java.util.HashMap;
 import java.util.Map;
-import model.player.Player;
-import model.player.Players;
+import model.participants.player.Player;
+import model.participants.player.Players;
 
 public class Profit {
     private final int dealerProfit;
