+++ src/main/java/model/blackjackgame/Result.java
@@ -4,9 +4,9 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
-import model.dealer.Dealer;
-import model.player.Player;
-import model.player.Players;
+import model.participants.dealer.Dealer;
+import model.participants.player.Player;
+import model.participants.player.Players;
 
 public class Result {
