+++ src/main/java/model/participants/player/Players.java
@@ -1,4 +1,4 @@
-package model.player;
+package model.participants.player;
 
 import static java.util.stream.Collectors.collectingAndThen;
 import static java.util.stream.Collectors.toList;
